Add where-is (closes feature request #1088394).
Remove list-functions (not in Emacs, and no longer needed).

Add case for space key to chordtostr.

Change some int's to size_t's in key conversion routines.

Fix a bug in write_bindings left over from conversion from alist to
list.

+++ src/extern.h
@@ -117,7 +117,7 @@ const char *next_history_element(History *hp);
 
 /* keys.c ----------------------------------------------------------------- */
 astr chordtostr(size_t key);
-int strtochord(char *buf, size_t *len);
+size_t strtochord(char *buf, size_t *len);
 int keystrtovec(char *key, size_t **keyvec);
 astr keyvectostr(size_t *keys, size_t numkeys);
 astr simplify_key(char *key);
